Rework main.rs to be compatible with clap 3.0.0 release candidate
Between the recent 3.0.0beta releases and the 3.0.0rc, clap made
a number of deprecations and enforced a number of backward
incompatible changes.

These changes to Cargo.toml and main.rs update the argument
parsing code to work in a manner consistent and compatible with
the new clap style.  Hopefully, there will not be any more
deprecations or backward-incompatible changes before the
final 3.0.0 release of clap.