feat(resolution): add USA support to parsimony and parsimony-em
This commit is the initital implementation of USA mode
support for the parsimony and parsimony-em resolution modes.
The output of this is NOT yet tested thoroughly. Currently, when
parsimony or parsimony-em are used in USA mode, a message noting
that this support is experimental is printed to the log. This
note should be removed once we are satisfied with the
implementation.