feat(resolution): Improve parsimony in USA mode
This commit makes the parsimony resolution algorithm
deterministic over runs.  Previously, there was a non-determinism
due to the HashSet having a random soruce that changed the
seed between runs.  Now, the hasher is set to be seeded
based on the barcode of each cell.  This produces
deterministic count output, while avoiding the potential
systematic bias of using the exact same seed for all cells.