feat(resolution): Add parsimony-gene and parsimony-gene-em modes
This code expands the list of modes that can be used for UMI resolution.
It adds a parsimony-gene and parsimony-gene-em mode.  These mirror the
existing parsimony and parsimony-em modes, with the exception that
the PUG on which the UMIs are resolved is defined in terms of
gene-level equivalence classes rather than transcript-level equivalence
classes.  This method trades off potential precision in the resolution
for the possibility of more robustness with respect to incomplete
annotation of the underlying transcripts.