fix(resolution): Fix indexing in alternative resolution for usa parsimony
When a large component is encountered, it is resolved with
`get_num_molecules_large_component`.  When in USA mode, the `num_genes`
parameter to this function should not be the number of genes
*considering* splicing status.  That is, in USA mode, this parameter
should be doubled.  Failing to do this can result in an out of bounds
access in the `counts` vec.  This commit fixes this issue by passing
the correct value for the number of gene targets based on the status
of USA mode.