refactor(cmd_interface): make mtx the deafult ouput
This commit makes the matrix market format (mtx) the
default output format of the quant and infer commands.
It introduces a new flag `--use-eds`, which is mutually
exclusive with `--use-mtx`, that can be used to ouput
in EDS format.  Note, the `--use-mtx` flag remains valid
so as to not break existing code that uses it, but it
now has essentially no effect since it is the default
behavior.

The new behavior changes the meaning of an alevin-fry
quant or infer invocation which contains neither
`--use-mtx` nor the new `--use-eds`. Previously,
such a command would write output in EDS format. Now, it
will write output in mtx format.