refactor(cmd_interface)!: remove --pug-exact-umi add --umi-edit-dist
Add a (hidden) --umi-edit-dist option that take a value informing the
underlying resolution algorithm of which edit distances to consider for
collapse among potentially colliding UMIs. Right now, 0 works with all
methods, while 1 only works with parsimony(-gene) and
parimony(-gene)-em. The default remain as before, 1 for parsimony(-gene)
and parsimony(-gene)-em, and 0 for all other methods. If the user
attempts to set an unsupported edit distance for a method, the
program will complain and exit.

BREAKING CHANGE: The hidden --pug-exact-umi flag (though never in a
released version of alevin-fry) is removed, and replaced with the more
general --umi-edit-dist.