refactor: remove duplicate name for parsimony-em
To streamline, simplify and robustify parsing, we have
removed the redundant name "full" that could previously
be passed as a resolution mode.  This mode was alway
simply an alias for "parsimony-em". As more resolution
modes have been and are being added, the name "full" for
this mode no longer really makes sense.  Further, "parsimony-em"
is more explicit and a better name for this mode. Moving
forward, "full" will no longer work as an alias for this mode
and will appear as unrecognized.